refactor: PlaceDetailResponseDto, KorDetailService 응답 구조 통일
